# src/TeamRotator.Api/swagger.yaml
openapi: 3.0.3
info:
  title: TeamRotator API
  version: 1.0.0
  description: API for managing team task rotations and assignments.

paths:
  /assignments:
    get:
      tags:
        - assignments
      summary: Retrieve the current rotation list
      description: Return a list of task assignments
      operationId: getRotationList
      responses:
        '200':
          description: Successful retrieval of rotation list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskAssignmentDto'
        '500':
          description: Internal server error

  /assignments/{id}:
    put:
      tags:
        - assignments
      summary: Update a task assignment by ID
      description: Modify an existing task assignment with provided info.
      operationId: updateRotationList
      parameters:
        - name: id
          in: path
          description: ID of the task assignment need to be updated
          required: true
          schema:
            type: integer
      requestBody:
        description: Assignment info needed
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModifyAssignmentDto'
      responses:
        '200':
          description: Successful update the task assignment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskAssignment'
        '400':
          description: Invalid ID or input data
        '404':
          description: Assignment not found
        '500':
          description: Internal server error

components:
  schemas:
    TaskAssignmentDto:
      type: object
      properties:
        id:
          type: integer
          description: Identifier for the task assignment
        taskId:
          type: integer
          description: ID of the task
        taskName:
          type: string
          description: Name of the task
        memberId:
          type: integer
          description: ID of the member
        host:
          type: string
          description: Host name of the member
        slackId:
          type: string
          description: Slack ID of the member
    TaskAssignment:
      type: object
      properties:
        id:
          type: integer
          description: Identifier for the task assignment
        taskId:
          type: integer
          description: ID of the task
        memberId:
          type: integer
          description: ID of the member
        startDate:
          type: string
          format: date
          description: Start date of the assignment
        endDate:
          type: string
          format: date
          description: End date of the assignment
    ModifyAssignmentDto:
      type: object
      required:
        - host
      properties:
        host:
          type: string
          description: Host name of the member to be assigned 